Use core mask for rknn
